deleted a few lines in main IJCSS nothing major

+++ IJCSS_editing/main.py
@@ -32,6 +32,8 @@ issue = volumes_path.split(r'/')[-1].split('Ed')[1][0:1]
 submit_folder = os.path.join(volumes_path, r'3-Uploads\10516-Volume{}-Issue{}'.format(volume,issue), '')
 if not os.path.isdir(submit_folder):
     os.mkdir(submit_folder)
+    
+a = 1
 
 for path, subdirs, files in os.walk(word_path):
     for name in files:
@@ -54,19 +56,19 @@ for path, subdirs, files in os.walk(word_path):
             submit_file_path_final = os.path.join(submit_folder,name.replace(file_extension,'.pdf'))
             
             # convert to PDF
-            if not os.path.isfile(pdf_file_path):
+            if not os.path.isfile(pdf_file_path) and a == 2:
                 convert(word_file_path, pdf_file_path) 
             
             # convert to post script
-            if not os.path.isfile(ps_file_path):
+            if not os.path.isfile(ps_file_path) and a == 2:
                 subprocess.call(['pdf2ps', pdf_file_path, ps_file_path])          
             
             # convert to pdf again (final)
-            if not os.path.isfile(pdf_file_path_final):
+            if not os.path.isfile(pdf_file_path_final) and a == 2:
                 subprocess.call(['ps2pdf', ps_file_path, pdf_file_path_final])
             
             # copy pdf to the final location 
-            if not os.path.isfile(submit_file_path_final):   
+            if not os.path.isfile(submit_file_path_final) and a == 2:
                 shutil.copyfile(pdf_file_path_final, submit_file_path_final)
